move locateMeButton under map
